added first code blocks

+++ src/simu/model/Asiakas.java
@@ -2,15 +2,14 @@ package simu.model;
 
 import simu.framework.*;
 
-// TODO:
-// Asiakas koodataan simulointimallin edellyttämällä tavalla (data!)
+
 public class Asiakas {
 	private double saapumisaika;
 	private double poistumisaika;
 	private int id;
 	private static int i = 1;
-	private static long sum = 0;
-	
+	private static double sum = 0;
+
 	public Asiakas(){
 	    id = i++;
